Move PostCard.tsx component into `post_card/`
